Update Home_Alarm_Device_Type
Minor code clean up and added code to alert user if AD2Pi is having difficulty communicating with the alarm panel.  In this case, the message !Sending....done will have 4 or more dots.  If 4 or more dots are detected, a message "Having Trouble Sending" will be fired off to the mobile device.  This can happen if the AD2Pi is not set with a valid keypad address for the alarm system.