Update Home_Alarm_Device_Type
An update to SmartThings device-type environment affected the input of the alarm code and created issues with both the input of the alarm code and the input of the panic key.  This update fixes the issue with input of the alarm code by specifying that the input is text.  Next update will fix the input of the panic key code.