Update and rename AD2SmartThings_v4_1.ino to AD2SmartThings_v4_2.ino
Reordered string passed to ST by putting keypadMsg at the end.  Added a length check since messages above 63 characters may not make it to ST and if greater than 63, truncate the message.  With keypadMsg at the end, the message will be truncated, but active/inactive zones will still be passed.